[refactor] move component to be under account folder
